Remove BigDecimal and use Long instead

+++ src/main/kotlin/com/ampnet/crowdfundingbackend/controller/pojo/response/ProjectResponse.kt
@@ -2,7 +2,6 @@ package com.ampnet.crowdfundingbackend.controller.pojo.response
 
 import com.ampnet.crowdfundingbackend.enums.Currency
 import com.ampnet.crowdfundingbackend.persistence.model.Project
-import java.math.BigDecimal
 import java.time.ZonedDateTime
 
 data class ProjectResponse(
@@ -14,18 +13,18 @@ data class ProjectResponse(
     val returnToInvestment: String,
     val startDate: ZonedDateTime,
     val endDate: ZonedDateTime,
-    val expectedFunding: BigDecimal,
+    val expectedFunding: Long,
     val currency: Currency,
-    val minPerUser: BigDecimal,
-    val maxPerUser: BigDecimal,
+    val minPerUser: Long,
+    val maxPerUser: Long,
     val mainImage: String?,
     val gallery: List<String>,
     val active: Boolean,
     val organization: OrganizationSmallResponse,
     val createByUser: String,
-    val currentFunding: BigDecimal
+    val currentFunding: Long
 ) {
-    constructor(project: Project, currentFunding: BigDecimal): this(
+    constructor(project: Project, currentFunding: Long): this(
             project.id,
             project.name,
             project.description,
